Stack memory optimisation in ssl_client2.c
For on-target system testing ssl_client2 has to be optimised for
the target devices that have limited memory. This commit does a basic
stack optimisation by moving the stack variables outside the main()
and declaring them static. As the variables are moved to bss, it
considerably reduces the stack usage and makes more space for the TLS stack.