Edit PKCS7 verify calls and test data generation
This commit is to put an end to the tedious CI failures on Windows. The issue was that, when genrating test files, Linux ends files with 0x0a while Windows ends with 0x0d0a. This was leading to a diffeence in generated hashes and pkcs7 signature verifiacation failing. So this commit, hard codes the binary to echo'd into the test file. Addtionally, I changed the calling of mbedtls_pk_verify because we were using the wrong value for hash length.

Signed-off-by: Nick Child <nick.child@ibm.com>