Open PKCS7 data files in binary format
Previously, we were reading files in text mode. This was causing all of the 0x0d0a line breaks to be converted to unix line breaks. This will cause issues when comparing hashes of files. If we read the files in binary mode (fopen "rb") then we avoid this conversion.

Signed-off-by: Nick Child <nick.child@ibm.com>